Pulling php out of collection reprts
